Moved io related modules from FOX.functions to FOX.io
