+++ FOX/README.rst
@@ -8,12 +8,6 @@ classes_
 
 Various classes implemented in **Auto-FOX**.
 
-~~~~~~~~~~
-functions_
-~~~~~~~~~~
-
-Various functions implemented in **Auto-FOX**.
-
 ~~~~~
 data_
 ~~~~~
@@ -26,8 +20,21 @@ examples_
 
 Various inputfile examples.
 
+~~~~~~~~~~
+functions_
+~~~~~~~~~~
+
+Various functions implemented in **Auto-FOX**.
+
+~~~
+io_
+~~~~
+
+Various io-related functions implemented in **Auto-FOX**.
+
 
 .. _classes: https://github.com/nlesc-nano/auto-FOX/tree/master/FOX/classes
 .. _functions: https://github.com/nlesc-nano/auto-FOX/tree/master/FOX/functions
 .. _data: https://github.com/nlesc-nano/auto-FOX/tree/master/FOX/data
 .. _examples: https://github.com/nlesc-nano/auto-FOX/tree/master/FOX/examples
+.. _io: https://github.com/nlesc-nano/auto-FOX/tree/master/FOX/io
