+++ FOX/functions/README.rst
@@ -26,29 +26,11 @@ cp2k_utils.py_
 
 A module with miscellaneous functions related to CP2K_.
 
-~~~~~~~~~~~~~~
-hdf5_utils.py_
-~~~~~~~~~~~~~~
-
-Functions for storing Monte Carlo results in HDF5_ format.
-
 ~~~~~~~
 rdf.py_
 ~~~~~~~
 
 A module for constructing radial distribution functions.
-
-~~~~~~~~~~~~
-read_prm.py_
-~~~~~~~~~~~~
-
-A module for reading and creating CHARMM_ parameter files (prm_).
-
-~~~~~~~~~~~~
-read_xyz.py_
-~~~~~~~~~~~~
-
-A module for reading multi-xyz files.
 
 ~~~~~~~~~
 utils.py_
@@ -61,13 +43,7 @@ A module with miscellaneous functions.
 .. _armc_sanitization: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/functions/armc_sanitization.py
 .. _charge_utils: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/functions/charge_utils.py
 .. _cp2k_utils: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/functions/cp2k_utils.py
-.. _hdf5_utils: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/functions/hdf5_utils.py
 .. _rdf: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/functions/rdf.py
-.. _read_xyz: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/functions/read_xyz.py
-.. _read_prm: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/functions/read_prm.py
 .. _utils: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/functions/utils.py
 
 .. _CP2K: https://www.cp2k.org/
-.. _HDF5: https://www.h5py.org/
-.. _CHARM: https://www.charmm.org/
-.. _prm: https://mackerell.umaryland.edu/charmm_ff.shtml
