+++ FOX/io/README.rst
@@ -1,0 +1,30 @@
+###
+io
+###
+
+~~~~~~~~~~~~~~
+hdf5_utils.py_
+~~~~~~~~~~~~~~
+
+Functions for storing Monte Carlo results in HDF5_ format.
+
+~~~~~~~~~~~~
+read_prm.py_
+~~~~~~~~~~~~
+
+A module for reading and creating CHARMM_ parameter files (prm_).
+
+~~~~~~~~~~~~
+read_xyz.py_
+~~~~~~~~~~~~
+
+A module for reading multi-xyz files.
+
+
+.. _hdf5_utils: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/io/hdf5_utils.py
+.. _read_xyz: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/io/read_xyz.py
+.. _read_prm: https://github.com/nlesc-nano/auto-FOX/blob/master/FOX/io/read_prm.py
+
+.. _HDF5: https://www.h5py.org/
+.. _CHARM: https://www.charmm.org/
+.. _prm: https://mackerell.umaryland.edu/charmm_ff.shtml
