+++ docs/5_xyz_reader.rst
@@ -4,10 +4,10 @@ Multi-XYZ reader
 ================
 
 A reader of multi-xyz files has been implemented in the
-:mod:`FOX.functions.read_xyz` module. The .xyz fileformat is designed
+:mod:`FOX.io.read_xyz` module. The .xyz fileformat is designed
 for storing the atomic symbols and cartesian coordinates of one or more
 molecules. The herein implemented
-:func:`FOX.functions.read_xyz.read_multi_xyz`
+:func:`FOX.io.read_xyz.read_multi_xyz`
 function allows for the fast, and memory-effiecient, retrieval of the
 various molecular geometries stored in an .xyz file.
 
@@ -27,4 +27,4 @@ a :class:`.MultiMolecule` instance.
 API
 ---
 
-.. autofunction:: FOX.functions.read_xyz.read_multi_xyz
+.. autofunction:: FOX.io.read_xyz.read_multi_xyz
